+++ src/type-system.ts
@@ -94,6 +94,11 @@ export namespace ElysiaTypeOptions {
 			| 'application/ld+json'
 			| 'application/xml'
 			| 'application/zip'
+			| 'font'
+			| 'font/woff2'
+			| 'font/woff'
+			| 'font/ttf'
+			| 'font/otf'
 		>
 		minSize?: FileUnit
 		maxSize?: FileUnit